Move some functions from util.py into type_util.py (#670)
* Mode some functions from util.py into type_util.py

* Lint
