+++ lib/streamlit/hashing.py
@@ -32,6 +32,7 @@ import textwrap
 import streamlit as st
 from streamlit import config
 from streamlit import util
+from streamlit import type_util
 from streamlit.folder_black_list import FolderBlackList
 from streamlit.compatibility import setup_2_3_shims
 
@@ -63,7 +64,7 @@ Context = namedtuple("Context", ["globals", "cells", "varnames"])
 
 
 def _is_magicmock(obj):
-    return util.is_type(obj, "unittest.mock.MagicMock") or util.is_type(
+    return type_util.is_type(obj, "unittest.mock.MagicMock") or type_util.is_type(
         obj, "mock.mock.MagicMock"
     )
 
@@ -133,8 +134,8 @@ def _key(obj, context):
             return ("__l", tuple(obj))
 
     if (
-        util.is_type(obj, "pandas.core.frame.DataFrame")
-        or util.is_type(obj, "numpy.ndarray")
+        type_util.is_type(obj, "pandas.core.frame.DataFrame")
+        or type_util.is_type(obj, "numpy.ndarray")
         or inspect.isbuiltin(obj)
         or inspect.isroutine(obj)
         or inspect.iscode(obj)
@@ -273,9 +274,9 @@ class CodeHasher:
                 return b"bool:1"
             elif obj is False:
                 return b"bool:0"
-            elif util.is_type(obj, "pandas.core.frame.DataFrame") or util.is_type(
-                obj, "pandas.core.series.Series"
-            ):
+            elif type_util.is_type(
+                obj, "pandas.core.frame.DataFrame"
+            ) or type_util.is_type(obj, "pandas.core.series.Series"):
                 import pandas as pd
 
                 if len(obj) >= PANDAS_ROWS_LARGE:
@@ -286,7 +287,7 @@ class CodeHasher:
                     # Use pickle if pandas cannot hash the object for example if
                     # it contains unhashable objects.
                     return pickle.dumps(obj, pickle.HIGHEST_PROTOCOL)
-            elif util.is_type(obj, "numpy.ndarray"):
+            elif type_util.is_type(obj, "numpy.ndarray"):
                 h = hashlib.new(self.name)
                 self._update(h, obj.shape)
 
@@ -302,8 +303,10 @@ class CodeHasher:
                 return self.to_bytes(obj.__name__)
             elif hasattr(obj, "name") and (
                 isinstance(obj, io.IOBase)
-                or (isinstance(obj.name, string_types)  # noqa: F821
-                    and os.path.exists(obj.name))
+                or (
+                    isinstance(obj.name, string_types)  # noqa: F821
+                    and os.path.exists(obj.name)
+                )
             ):
                 # Hash files as name + last modification date + offset.
                 h = hashlib.new(self.name)
