+++ lib/streamlit/util.py
@@ -59,7 +59,8 @@ GITBLOB_RE = re.compile(
     "(?P<base>https:\/\/?(gist.)?github.com\/)"
     "(?P<account>([\w\.]+\/){1,2})"
     "(?P<blob_or_raw>(blob|raw))?"
-    "(?P<suffix>(.+)?)")
+    "(?P<suffix>(.+)?)"
+)
 
 
 def _decode_ascii(string):
@@ -284,36 +285,6 @@ def open_browser(url):
         subprocess.Popen(cmd, stdout=devnull, stderr=subprocess.STDOUT)
 
 
-def is_type(obj, fqn_type_pattern):
-    """Check type without importing expensive modules.
-
-    Parameters
-    ----------
-    obj : any
-        The object to type-check.
-    fqn_type_pattern : str or regex
-        The fully-qualified type string or a regular expression.
-        Regexes should start with `^` and end with `$`.
-
-    Example
-    -------
-
-    To check whether something is a Matplotlib Figure without importing
-    matplotlib, use:
-
-    >>> is_type(foo, 'matplotlib.figure.Figure')
-
-    """
-    the_type = type(obj)
-    module = the_type.__module__
-    name = the_type.__name__
-    actual_fqn = "%s.%s" % (module, name)
-    if isinstance(fqn_type_pattern, string_types):
-        return fqn_type_pattern == actual_fqn
-    else:
-        return fqn_type_pattern.match(actual_fqn) is not None
-
-
 class Error(Exception):
     pass
 
@@ -326,82 +297,6 @@ def is_pex():
     """
     if re.match(r".*pex$", sys.path[0]):
         return True
-    return False
-
-
-def is_sympy_expession(obj):
-    """True if input is a SymPy expression."""
-    if not is_type(obj, re.compile(r"^sympy.*$")):
-        return False
-
-    try:
-        import sympy
-
-        if isinstance(obj, sympy.Expr):
-            return True
-    except:
-        return False
-
-
-def is_altair_chart(obj):
-    """True if input looks like an Altair chart."""
-    return is_type(obj, re.compile(r"^altair\.vegalite\.v\d+\.api\.\w*Chart$"))
-
-
-def is_keras_model(obj):
-    """True if input looks like a Keras model."""
-    return (
-        is_type(obj, "keras.engine.sequential.Sequential")
-        or is_type(obj, "keras.engine.training.Model")
-        or is_type(obj, "tensorflow.python.keras.engine.sequential.Sequential")
-        or is_type(obj, "tensorflow.python.keras.engine.training.Model")
-    )
-
-
-def is_plotly_chart(obj):
-    """True if input looks like a Plotly chart."""
-    return (
-        is_type(obj, "plotly.graph_objs._figure.Figure")
-        or _is_list_of_plotly_objs(obj)
-        or _is_probably_plotly_dict(obj)
-    )
-
-
-def is_graphviz_chart(obj):
-    """True if input looks like a GraphViz chart."""
-    return is_type(obj, "graphviz.dot.Graph") or is_type(obj, "graphviz.dot.Digraph")
-
-
-def _is_plotly_obj(obj):
-    """True if input if from a type that lives in plotly.plotly_objs."""
-    the_type = type(obj)
-    return the_type.__module__.startswith("plotly.graph_objs")
-
-
-def _is_list_of_plotly_objs(obj):
-    if type(obj) is not list:
-        return False
-    if len(obj) == 0:
-        return False
-    return all(_is_plotly_obj(item) for item in obj)
-
-
-def _is_probably_plotly_dict(obj):
-    if type(obj) not in dict_types:
-        return False
-
-    if len(obj.keys()) == 0:
-        return False
-
-    if any(k not in ["config", "data", "frames", "layout"] for k in obj.keys()):
-        return False
-
-    if any(_is_plotly_obj(v) for v in obj.values()):
-        return True
-
-    if any(_is_list_of_plotly_objs(v) for v in obj.values()):
-        return True
-
     return False
 
 
@@ -421,14 +316,6 @@ def is_repl():
         return True
 
     return False
-
-
-_function_type = type(lambda: 0)
-
-
-def is_function(x):
-    """Return True if x is a function."""
-    return type(x) == _function_type
 
 
 def get_streamlit_file_path(*filepath):
@@ -459,9 +346,9 @@ def print_url(title, url):
 
 
 def process_gitblob_url(url):
-    """Check url to see if it describes a GitHub Gist "blob" URL.  
+    """Check url to see if it describes a GitHub Gist "blob" URL.
 
-    If so, returns a new URL to get the "raw" script. 
+    If so, returns a new URL to get the "raw" script.
     If not, returns URL unchanged.
     """
     # Matches github.com and gist.github.com.  Will not match githubusercontent.com.
@@ -492,17 +379,6 @@ def get_hostname(url):
 
     parsed = urllib.parse.urlparse(url)
     return parsed.hostname
-
-
-def is_namedtuple(x):
-    t = type(x)
-    b = t.__bases__
-    if len(b) != 1 or b[0] != tuple:
-        return False
-    f = getattr(t, "_fields", None)
-    if not isinstance(f, tuple):
-        return False
-    return all(type(n).__name__ == "str" for n in f)
 
 
 def is_darwin():
