+++ lib/tests/streamlit/util_test.py
@@ -13,8 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Config System Unittest."""
-
 # Python 2/3 compatibility
 from __future__ import print_function, division, unicode_literals, absolute_import
 
@@ -122,58 +120,6 @@ class UtilTest(unittest.TestCase):
             )
             self.assertEqual(str(e.value), error_msg)
 
-    def test_list_is_plotly_chart(self):
-        trace0 = go.Scatter(x=[1, 2, 3, 4], y=[10, 15, 13, 17])
-        trace1 = go.Scatter(x=[1, 2, 3, 4], y=[16, 5, 11, 9])
-        data = [trace0, trace1]
-
-        res = util.is_plotly_chart(data)
-        self.assertTrue(res)
-
-    def test_data_dict_is_plotly_chart(self):
-        trace0 = go.Scatter(x=[1, 2, 3, 4], y=[10, 15, 13, 17])
-        trace1 = go.Scatter(x=[1, 2, 3, 4], y=[16, 5, 11, 9])
-        d = {"data": [trace0, trace1]}
-
-        res = util.is_plotly_chart(d)
-        self.assertTrue(res)
-
-    def test_dirty_data_dict_is_not_plotly_chart(self):
-        trace0 = go.Scatter(x=[1, 2, 3, 4], y=[10, 15, 13, 17])
-        trace1 = go.Scatter(x=[1, 2, 3, 4], y=[16, 5, 11, 9])
-        d = {"data": [trace0, trace1], "foo": "bar"}  # Illegal property!
-
-        res = util.is_plotly_chart(d)
-        self.assertFalse(res)
-
-    def test_layout_dict_is_not_plotly_chart(self):
-        d = {
-            # Missing a component with a graph object!
-            "layout": {"width": 1000}
-        }
-
-        res = util.is_plotly_chart(d)
-        self.assertFalse(res)
-
-    def test_fig_is_plotly_chart(self):
-        trace1 = go.Scatter(x=[1, 2, 3, 4], y=[16, 5, 11, 9])
-
-        # Plotly 3.7 needs to read the config file at /home/.plotly when
-        # creating an image. So let's mock that part of the Figure creation:
-        with patch("plotly.offline.offline._get_jconfig") as mock:
-            mock.return_value = {}
-            fig = go.Figure(data=[trace1])
-
-        res = util.is_plotly_chart(fig)
-        self.assertTrue(res)
-
-    def test_is_namedtuple(self):
-        Boy = namedtuple("Boy", ("name", "age"))
-        John = Boy("John", "29")
-
-        res = util.is_namedtuple(John)
-        self.assertTrue(res)
-
     def test_get_external_ip(self):
         # Test success
         with requests_mock.mock() as m:
@@ -200,8 +146,7 @@ class UtilTest(unittest.TestCase):
 
 
 class FileIsInFolderTest(unittest.TestCase):
-    """Tests for file_is_in_folder.
-    """
+    """Tests for file_is_in_folder."""
 
     def test_file_in_folder(self):
         # Test with and without trailing slash
@@ -249,30 +194,49 @@ class FileIsInFolderTest(unittest.TestCase):
 
 
 class GitHubUrlTest(unittest.TestCase):
-    GITHUB_URLS = [('https://github.com/aritropaul/streamlit/blob/b72adbcf00c91775db14e739e2ea33d6df9079c3/lib/streamlit/cli.py', 
-                    'https://github.com/aritropaul/streamlit/raw/b72adbcf00c91775db14e739e2ea33d6df9079c3/lib/streamlit/cli.py'),
-                   ('https://github.com/streamlit/streamlit/blob/develop/examples/video.py',
-                    'https://github.com/streamlit/streamlit/raw/develop/examples/video.py'),
-                   ('https://github.com/text2gene/text2gene/blob/master/sbin/clinvar.hgvs_citations.sql',
-                    'https://github.com/text2gene/text2gene/raw/master/sbin/clinvar.hgvs_citations.sql'),
-                   ('https://github.com/mekarpeles/math.mx/blob/master/README.md',
-                    'https://github.com/mekarpeles/math.mx/raw/master/README.md'),
-                  ]
+    GITHUB_URLS = [
+        (
+            "https://github.com/aritropaul/streamlit/blob/b72adbcf00c91775db14e739e2ea33d6df9079c3/lib/streamlit/cli.py",
+            "https://github.com/aritropaul/streamlit/raw/b72adbcf00c91775db14e739e2ea33d6df9079c3/lib/streamlit/cli.py",
+        ),
+        (
+            "https://github.com/streamlit/streamlit/blob/develop/examples/video.py",
+            "https://github.com/streamlit/streamlit/raw/develop/examples/video.py",
+        ),
+        (
+            "https://github.com/text2gene/text2gene/blob/master/sbin/clinvar.hgvs_citations.sql",
+            "https://github.com/text2gene/text2gene/raw/master/sbin/clinvar.hgvs_citations.sql",
+        ),
+        (
+            "https://github.com/mekarpeles/math.mx/blob/master/README.md",
+            "https://github.com/mekarpeles/math.mx/raw/master/README.md",
+        ),
+    ]
 
-    GIST_URLS = [('https://gist.github.com/nthmost/b521b80fbd834e67b3f5e271e9548232', 
-                  'https://gist.github.com/nthmost/b521b80fbd834e67b3f5e271e9548232/raw'),
-                 ('https://gist.github.com/scottyallen/1888e058261fc21f184f6be192bbe131', 
-                  'https://gist.github.com/scottyallen/1888e058261fc21f184f6be192bbe131/raw'),
-                 ('https://gist.github.com/tvst/faf057abbedaccaa70b48216a1866cdd',
-                  'https://gist.github.com/tvst/faf057abbedaccaa70b48216a1866cdd/raw'),
-                ]
+    GIST_URLS = [
+        (
+            "https://gist.github.com/nthmost/b521b80fbd834e67b3f5e271e9548232",
+            "https://gist.github.com/nthmost/b521b80fbd834e67b3f5e271e9548232/raw",
+        ),
+        (
+            "https://gist.github.com/scottyallen/1888e058261fc21f184f6be192bbe131",
+            "https://gist.github.com/scottyallen/1888e058261fc21f184f6be192bbe131/raw",
+        ),
+        (
+            "https://gist.github.com/tvst/faf057abbedaccaa70b48216a1866cdd",
+            "https://gist.github.com/tvst/faf057abbedaccaa70b48216a1866cdd/raw",
+        ),
+    ]
 
-    INVALID_URLS = ['blah', 'google.com', 'http://homestarrunner.com',
-                    'https://somethinglikegithub.com/withablob',
-                    'gist.github.com/nothing',
-                    'https://raw.githubusercontent.com/streamlit/streamlit/develop/examples/video.py',
-                    'streamlit.io/raw/blob',
-                   ]
+    INVALID_URLS = [
+        "blah",
+        "google.com",
+        "http://homestarrunner.com",
+        "https://somethinglikegithub.com/withablob",
+        "gist.github.com/nothing",
+        "https://raw.githubusercontent.com/streamlit/streamlit/develop/examples/video.py",
+        "streamlit.io/raw/blob",
+    ]
 
     def test_github_url_is_replaced(self):
         for (target, processed) in self.GITHUB_URLS:
@@ -285,4 +249,3 @@ class GitHubUrlTest(unittest.TestCase):
     def test_nonmatching_url_is_not_replaced(self):
         for url in self.INVALID_URLS:
             assert url == util.process_gitblob_url(url)
-
